Implement special case keys for making sure you get the same server using distredis:
sometimes you want to make sure a set of keys reside on the same server
when using DistRedis, so I;ve implemented a scheme for this based on Salvatore's
idea. if you use a key with {} embeded distredis will hash the key based
on the string between the {} so for example:

foo{users}
bar{users}
urmom{users}

all of these will hash to the same server as the bare key users. This allows
you to enforce a server in a distributed env in case you want to do set intersection or
whatever and need the same server.