Revert to using buffered write (fixes #108)
It was changed to `syswrite` in dcb249 because of slightly better
throughput. This appeared to do partial writes on OSX's default Ruby
(1.8.7-p174) with payload size close to the maximum write buffer. This
behavior doesn't show with later patch versions of Ruby. Reverting to
buffered write because the extra overhead of checking the number of
bytes written in a partial write is not worth it...