Add an `inherit_socket` option to disable fork safety check
When `inherit_socket` is set to true, we disable safety check that prevents
a forked child from sharing a socket with its parent; this is potentially
useful in order to prohibit excessive connection churn on redis when ALL of the following are met:

 - many short-lived forked children of one process need to talk to redis, AND
 - your own code prevents the parent process from using the redis connection
   while a child is alive

Improper use of this option will result in corrupted and/or incorrect responses.

`inherit_socket` is disabled by default.