Initial native support for Sentinel.
This commit implements native support for Sentinel. At API level it is
exposed using the :url connection method, where the master name is
passed using the sentinel://mastername scheme. It is possible to pass
a database number as well adding an additional slash.

When Sentinel support is used, it is mandatory to pass a list of
Senitnel instances to contact to get the configuration update. The list
is passed as an Array of Hashes with :host and :port.

By default it will try to resolve who is the current master, and will
connect to it. However we may want to connect to a slave. For this
:role => :slave should be passed as an additional option, but
this support is not implemented by this commit, so :role => :master is
hte only valid one.

This code implements the new Sentinel - Client interaction. See the
discussion in the Redis mailing list here:

https://groups.google.com/d/msg/redis-db/uNqWceEjV8g/wfoSy24ktJcJ

However the code is designed to be backward compatible with older Redis
and Sentinel instances not supporting, respectively, the ROLE command or
the disconnection of clients on role switch. Of course without the
new protocol the system is less safe.