Allow multiple arguments for `pfcount`
It now supports both an array of keys and multiple arguments.
For distributed access, it will now also ensure that all keys
are on the same node.

The tests were changed to do correctly work both for both scenarios
(keys were extended to use tags, so they land on the same node)
Additionally, a test for this behaviour was added.

Closes #450.