Fix failing jruby tests, change default REDIS_BRANCH to 4.0
jruby tests fail when using jruby_openssl 0.10.0 but not when using 0.9.21
This change adds a Gemfile dependency for jruby_openssl <0.10.0

This change also switches the default REDIS_BRANCH to 4.0 since
the unstable branch of redis isn't always in a buildable state.

Also adds minor tweaks to how the tests are run due to minor issues with Gemfile usage.